Move use-hidden-description from src/button-dropdown/utils to src/internal/hooks
