+++ src/internal/hooks/use-hidden-description/index.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import React from 'react';
 
-import { useUniqueId } from '../../internal/hooks/use-unique-id';
+import { useUniqueId } from '../use-unique-id';
 
 export default function useHiddenDescription(description?: string) {
   const id = useUniqueId();
